Exercise4
A bank intends to design a program to display the denomination of an input amount. 
The available denomination with the bank are 1000 , 500 , 100 , 50 , 20 , 10 , 5 , 2 , and 1.

Design a program to accept the amount from the user and display the break-up in descending 
order of denomination. (i.e. preference should be given to the highest denomination available)
along with the total number of notes. [Note: Only the denomination used, should be displayed].