Exercise3
Brandon Foster is an interplanetary space Boxer, who  currently holds the championship belts for various different categories on many different planets. However, it is often difficult for him to recall what his "target weight" needs to be on earth in order to make the weight class on other planets. Make a Java program to help him keep track of this.

It should ask him his weight on earth, and to enter a number for the planet he wants to fight on. It should then compute his weight on the destination planet based on the table below:

#	Planet	Relative gravity
1	Venus	0.78
2	Mars	0.39
3	Jupiter	2.65
4	Saturn	1.17
5	Uranus	1.05
6	Neptune	1.23

Note: You are to use following formula when calculating the weight on the given planet (weight on earth * relative gravity)