Exercise2
Class B extends Class A. Class A has two member variables 'i' and 'j', a constructor and a method show() which displays the value of i and j.
Class B has a member variable 'k', a constructor. Write a method show() in class B which overrides the show() method of class A and displays the value of k.