Exercise3
We generally use Inheritance to create an exception superclass and various exception subclasses. Given the classes ExceptionA, ExceptionB, and ExceptionC.
ExceptionA is the super class of ExceptionB.
ExceptionC is the subclass of ExceptionB.
Write the code to demonstrate that the catch specifying superclass ExceptionA, also catches subclass ExceptionB's exceptions too.