Exercise3
Write a program that creates an ArrayList which can hold Integers. Enter the number of integers that will be inserted, then enter those integers. Then use a linear search to find the largest value in the ArrayList, and display that value, and the slot number for it.

For example:

Input:
3
=====
1

2

3

Output:

The largest value is 3; which is in slot 2